Calc LL in float64.
